Put models into subapps
