+++ backend excercise/participation/quiz/views.py
@@ -1,0 +1,110 @@
+from venv import logger
+from django.http import Http404
+from django.shortcuts import get_object_or_404, render
+from rest_framework.exceptions import ValidationError
+from rest_framework.response import Response
+from rest_framework import status
+from base_view import BaseView
+from participation.quiz.models import Quiz
+from participation.permissions import AnswerPermission, IsAdminOrReadOnly, IsAdminOrReadUpdate
+from participation.quiz.serializers import QuizSerializer
+from users.permissions import IsAdmin, IsAdminOrTeacher, IsTeacher
+from drf_spectacular.utils import extend_schema, OpenApiParameter, OpenApiExample
+from drf_spectacular.types import OpenApiTypes
+from django.db.models import Sum, Count
+from quiz.models import Question
+from django.utils import timezone
+from drf_spectacular.utils import extend_schema, OpenApiParameter
+from drf_spectacular.types import OpenApiTypes
+# Create your views here.
+@extend_schema(tags=['Participation/ Quiz'])
+class QuizAPIView(BaseView):
+    serializer_class = QuizSerializer
+    permission_classes = [IsAdminOrReadOnly] 
+    
+    def get(self, request, pk=None):
+        try:
+            if pk:
+                quiz = get_object_or_404(Quiz, pk=pk)
+                serializer = QuizSerializer(quiz)
+                return self.send_successful_response(serializer.data)
+            quizzes = Quiz.objects.all()
+            serializer = QuizSerializer(quizzes, many=True)
+            return self.send_successful_response(serializer.data)
+        except Http404:
+            return self.send_bad_response(
+                {"detail": "Quiz not found."}, status_code=status.HTTP_404_NOT_FOUND
+            )
+        except Exception as e:
+            logger.exception("Unexpected error in GET Quiz:")
+            return self.send_bad_response(
+                {"detail": "An unexpected error occurred."},
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR
+            )
+
+    def post(self, request):
+        try:
+            serializer = QuizSerializer(data=request.data)
+            serializer.is_valid(raise_exception=True)
+            serializer.save()
+            return self.send_201_response(serializer.data)
+        except ValidationError as e:
+            return self.send_bad_response(e.detail, status_code=status.HTTP_400_BAD_REQUEST)
+        except Exception as e:
+            logger.exception("Unexpected error in POST Quiz:")
+            return self.send_bad_response(
+                {"detail": "An unexpected error occurred."},
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR
+            )
+
+    def put(self, request, pk):
+        try:
+            quiz = get_object_or_404(Quiz, pk=pk)
+            serializer = QuizSerializer(quiz, data=request.data)
+            serializer.is_valid(raise_exception=True)
+            serializer.save()
+            return self.send_successful_response(serializer.data)
+        except Http404:
+            return self.send_bad_response(
+                {"detail": "Quiz not found."}, status_code=status.HTTP_404_NOT_FOUND
+            )
+        except ValidationError as e:
+            return self.send_bad_response(e.detail, status_code=status.HTTP_400_BAD_REQUEST)
+        except Exception as e:
+            logger.exception("Unexpected error in PUT Quiz:")
+            return self.send_bad_response(
+                {"detail": "An unexpected error occurred."},
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR
+            )
+
+    def patch(self, request, pk):
+        try:
+            quiz = get_object_or_404(Quiz, pk=pk)
+            serializer = QuizSerializer(quiz, data=request.data, partial=True)
+            serializer.is_valid(raise_exception=True)
+            serializer.save()
+            return self.send_successful_response(serializer.data)
+        except Http404:
+            return self.send_bad_response(
+                {"detail": "Quiz not found."}, status_code=status.HTTP_404_NOT_FOUND
+            )
+        except ValidationError as e:
+            return self.send_bad_response(e.detail, status_code=status.HTTP_400_BAD_REQUEST)
+        except Exception as e:
+            logger.exception("Unexpected error in PATCH Quiz:")
+            return self.send_bad_response(
+                {"detail": "An unexpected error occurred."},
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR
+            )
+        
+    def delete(self, request, pk):
+            try:
+                quiz = get_object_or_404(Quiz, pk=pk)
+                quiz.delete()
+                return self.send_no_content_response()
+            except Http404:
+                return self.send_bad_response({"detail": "Quiz not found."}, status_code=status.HTTP_404_NOT_FOUND)
+            except Exception as e:
+                logger.exception("Unexpected error on DELETE Quiz:")
+                return self.send_bad_response({"detail": "An unexpected error occurred."}, status_code=status.HTTP_500_INTERNAL_SERVER_ERROR)
+
